Create SettlementTemplateConfg class Move Template logic from SettlementConfig to SettlementTemplateConfig Remove extends UserConfigurableConfig from SettlementConfig Change usage of getItem in all classes to SettlementTemplateConfig Change some while to for loop Remove incorrect @params
