+++ mars-sim-core/src/main/java/com/mars_sim/core/interplanetary/transport/resupply/ResupplyUtil.java
@@ -15,6 +15,7 @@ import com.mars_sim.core.UnitManager;
 import com.mars_sim.core.environment.MarsSurface;
 import com.mars_sim.core.structure.Settlement;
 import com.mars_sim.core.structure.SettlementConfig;
+import com.mars_sim.core.structure.SettlementTemplateConfig;
 import com.mars_sim.core.time.MarsTime;
 
 /**
@@ -44,7 +45,7 @@ public final class ResupplyUtil {
 	public static List<Resupply> loadInitialResupplyMissions(Simulation sim) {
 		MarsTime currentTime = sim.getMasterClock().getMarsTime();
 		UnitManager unitManager = sim.getUnitManager();
-		SettlementConfig settlementConfig = sim.getConfig().getSettlementConfiguration();
+		SettlementTemplateConfig settlementTemplateConfig = sim.getConfig().getSettlementTemplateConfiguration();
 
 		List<Resupply> resupplies = new ArrayList<>();
 			
@@ -52,7 +53,7 @@ public final class ResupplyUtil {
 			String templateName = settlement.getTemplate();
 
 			// For each Settlement get the resupply scheduled defined by the SettlementTemplate
-			for(ResupplySchedule template : settlementConfig.getItem(templateName).getResupplyMissionTemplates()) {
+			for(ResupplySchedule template : settlementTemplateConfig.getItem(templateName).getResupplyMissionTemplates()) {
 				var schedule = template.getSchedule();
 				// Get the local time at teh Settlemnt when this will arrive
 				MarsTime arrivalDate = schedule.getFirstEvent(currentTime,
